kafka/server/tests: Refactor - move create_topic to helpers
Pure refactor, no functional change.
