+++ src/v/kafka/server/tests/alter_config_test.cc
@@ -43,6 +43,7 @@ inline ss::logger test_log("test"); // NOLINT
 
 class alter_config_test_fixture : public topic_properties_test_fixture {
 public:
+    using topic_properties_test_fixture::create_topic;
     void create_topic(
       model::topic name,
       int partitions,
@@ -58,53 +59,6 @@ public:
                 model::ntp(tp_ns.ns, tp_ns.tp, model::partition_id(i)),
                 model::offset(0));
           })
-          .get();
-    }
-
-    template<typename Func>
-    auto do_with_client(Func&& f) {
-        return make_kafka_client().then(
-          [f = std::forward<Func>(f)](kafka::client::transport client) mutable {
-              return ss::do_with(
-                std::move(client),
-                [f = std::forward<Func>(f)](
-                  kafka::client::transport& client) mutable {
-                    return client.connect().then(
-                      [&client, f = std::forward<Func>(f)]() mutable {
-                          return f(client);
-                      });
-                });
-          });
-    }
-
-    kafka::create_topics_response create_topic(
-      const model::topic& tp,
-      const absl::flat_hash_map<ss::sstring, ss::sstring>& properties,
-      int num_partitions = 1,
-      int16_t replication_factor = 1) {
-        kafka::creatable_topic topic{
-          .name = model::topic(tp),
-          .num_partitions = num_partitions,
-          .replication_factor = replication_factor,
-        };
-        for (auto& [k, v] : properties) {
-            kafka::createable_topic_config config;
-            config.name = k;
-            config.value = v;
-            topic.configs.push_back(std::move(config));
-        }
-
-        auto req = kafka::create_topics_request{.data{
-          .topics = {topic},
-          .timeout_ms = 10s,
-          .validate_only = false,
-        }};
-
-        return do_with_client([req = std::move(req)](
-                                kafka::client::transport& client) mutable {
-                   return client.dispatch(
-                     std::move(req), kafka::api_version(0));
-               })
           .get();
     }
